Move operations, schema, models to store mod
This change moves the remaining operations, schema and models modules to
the store module.
