+++ libtransact/src/state/merkle/sql/store/models/sqlite.rs
@@ -25,7 +25,7 @@ use diesel::{
     sqlite::Sqlite,
 };
 
-use crate::state::merkle::sql::schema::sqlite_merkle_radix_tree_node;
+use crate::state::merkle::sql::store::schema::sqlite_merkle_radix_tree_node;
 
 #[derive(Insertable, Queryable, QueryableByName, Identifiable)]
 #[cfg_attr(test, derive(Debug, PartialEq))]
