+++ libtransact/src/state/merkle/sql/store/operations/has_root.rs
@@ -21,9 +21,9 @@ use diesel::{dsl::exists, select};
 
 use crate::error::InternalError;
 #[cfg(feature = "postgres")]
-use crate::state::merkle::sql::schema::postgres_merkle_radix_tree_node;
+use crate::state::merkle::sql::store::schema::postgres_merkle_radix_tree_node;
 #[cfg(feature = "sqlite")]
-use crate::state::merkle::sql::schema::sqlite_merkle_radix_tree_node;
+use crate::state::merkle::sql::store::schema::sqlite_merkle_radix_tree_node;
 
 use super::MerkleRadixOperations;
 
@@ -60,9 +60,11 @@ mod tests {
     use diesel::dsl::insert_into;
 
     #[cfg(feature = "state-merkle-sql-postgres-tests")]
-    use crate::state::merkle::sql::{backend::postgres::test::run_postgres_test, models::postgres};
+    use crate::state::merkle::sql::{
+        backend::postgres::test::run_postgres_test, store::models::postgres,
+    };
     #[cfg(feature = "sqlite")]
-    use crate::state::merkle::sql::{migration, models::sqlite};
+    use crate::state::merkle::sql::{migration, store::models::sqlite};
 
     /// Test that has_node succeeds if the root hash is in the tree table.
     #[cfg(feature = "sqlite")]
